:clown_face: ajout d'un mathalea.tex pour tests

+++ zip.php
@@ -38,7 +38,7 @@
         if (is_dir($thisdir) ) {
             $handle=opendir($thisdir);
             while (false!==($file = readdir($handle))) {
-                if ($file != "." && $file != ".." && $file != "automultiplechoice.sty" && $file != "Projet.zip" && $file != "ProjetCrashed.zip") { 
+                if ($file != "." && $file != ".." && $file != "automultiplechoice.sty" && $file != "Projet.zip" && $file != "ProjetCrashed.zip" && $file != "mathalea.tex") { 
                     $Diff = (time() - filectime("$thisdir/$file"));
                     if ($Diff > $sec) unlink("$thisdir/$file");
                 }
@@ -49,43 +49,46 @@
 
         if ($zip->open($thisdir.'/'.$filename, ZipArchive::CREATE)!==TRUE) {
             exit("Impossible d'ouvrir le fichier <$filename>\n");
-        }
+        } else {
 
-        // On nettoie la chaine du formulaire pour éviter les injections
-        $encoded_snip = $_POST['encoded_snip'];
-        $encoded_snip = trim($encoded_snip);
-        // $encoded_snip = stripslashes($encoded_snip);
-        // $encoded_snip = htmlspecialchars($encoded_snip);
+            $zip->open($thisdir.'/'.$filename, ZipArchive::CREATE);
+            // On nettoie la chaine du formulaire pour éviter les injections
+            $encoded_snip = $_POST['encoded_snip'];
+            $encoded_snip = trim($encoded_snip);
+            // $encoded_snip = stripslashes($encoded_snip);
+            // $encoded_snip = htmlspecialchars($encoded_snip);
 
-        $zip->addFromString("mathalea" . time().".tex", $encoded_snip);                   
-        $zip->addFile($thisdir."/automultiplechoice.sty","/automultiplechoice.sty");
-        echo "Nombre de fichiers : " . $zip->numFiles . "\n";
-        echo "Statut :" . $zip->status . "\n";
-        $zip->close();
-        if(file_exists($thisdir.'/'.$filename)){
-            echo "file exist";            
-            if (get_ip()=='127.0.0.1') { // Si on est en local , on lance la routine de téléchargement
-                header('Content-type: application/zip'); // on indique que c'est une archive
-                header('Content-Transfer-Encoding: fichier'); // transfert en binaire (fichier)
-                header('Content-Disposition: attachment; filename="'.$thisdir.'/'.$filename.'"'); // nom de l'archive
-                header('Content-Length: '.filesize($thisdir.'/'.$filename)); // taille de l'archive
-                header('Pragma: no-cache'); 
-                header('Expires: 0');
-                header("location:$thisdir/$filename"); // redirection vers le téléchargement de l'archive en local
-                //header("location:https://www.overleaf.com/docs?snip_uri=https://coopmaths.fr/fichiers/images.zip"); // pour tester avec l'existant
-            } else {
-                //$domaine = "https://coopmaths.fr";
-                $domaine = "http://mathalea.mathslozano.fr";
-                header('Content-type: application/zip'); // on indique que c'est une archive
-                header('Content-Transfer-Encoding: fichier'); // transfert en binaire (fichier)
-                header('Content-Disposition: attachment; filename="'.$thisdir.'/'.$filename.'"'); // nom de l'archive
-                header('Content-Length: '.filesize($thisdir.'/'.$filename)); // taille de l'archive
-                header('Pragma: no-cache'); 
-                header('Expires: 0');
-                header("location:https://www.overleaf.com/docs?snip_uri=$domaine/$thisdir/$filename"); // pour passer à overleaf
-                //header("location:https://www.overleaf.com/docs?snip_uri=https://coopmaths.fr/fichierszip/Projet.zip"); // pour tester avec l'existant
-            }
+            $zip->addFromString("mathalea" . time().".tex", $encoded_snip);                   
+            $zip->addFile($thisdir."/automultiplechoice.sty","/automultiplechoice.sty");
+            echo "Nombre de fichiers : " . $zip->numFiles . "\n";
+            echo "Statut :" . $zip->status . "\n";
+            $zip->close();
+            if(file_exists($thisdir.'/'.$filename)){
+                echo "file exist";            
+                if (get_ip()=='127.0.0.1') { // Si on est en local , on lance la routine de téléchargement
+                    header('Content-type: application/zip'); // on indique que c'est une archive
+                    header('Content-Transfer-Encoding: fichier'); // transfert en binaire (fichier)
+                    header('Content-Disposition: attachment; filename="'.$thisdir.'/'.$filename.'"'); // nom de l'archive
+                    header('Content-Length: '.filesize($thisdir.'/'.$filename)); // taille de l'archive
+                    header('Pragma: no-cache'); 
+                    header('Expires: 0');
+                    header("location:$thisdir/$filename"); // redirection vers le téléchargement de l'archive en local
+                    //header("location:https://www.overleaf.com/docs?snip_uri=https://coopmaths.fr/fichiers/images.zip"); // pour tester avec l'existant
+                } else {
+                    $domaine = "https://coopmaths.fr";
+                    //$domaine = "http://mathalea.mathslozano.fr";
+                    header('Content-type: application/zip'); // on indique que c'est une archive
+                    header('Content-Transfer-Encoding: fichier'); // transfert en binaire (fichier)
+                    header('Content-Disposition: attachment; filename="'.$thisdir.'/'.$filename.'"'); // nom de l'archive
+                    header('Content-Length: '.filesize($thisdir.'/'.$filename)); // taille de l'archive
+                    header('Pragma: no-cache'); 
+                    header('Expires: 0');
+                    header("location:https://www.overleaf.com/docs?snip_uri=$domaine/$thisdir/$filename"); // pour passer à overleaf
+                    //header("location:https://www.overleaf.com/docs?snip_uri=https://coopmaths.fr/fichierszip/Projet.zip"); // pour tester avec l'existant
+                }
+            };
         };
+
 
     } else {
         $error .= "KO";
